# Fix live reporting of latency