--- a/src/benchmarklib/benchmark_runner.cpp
+++ b/src/benchmarklib/benchmark_runner.cpp
@@ -274,11 +274,11 @@ void BenchmarkRunner::_benchmark_ordered() {
     boost::accumulators::accumulator_set<double, boost::accumulators::stats<boost::accumulators::tag::mean>>
         accumulator;
     for (const auto& entry : result.successful_runs) {
-      auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(entry.duration);
+      auto duration = std::chrono::duration_cast<std::chrono::nanoseconds>(entry.duration);
       accumulator(static_cast<double>(duration.count()));
     }
     auto mean_in_milliseconds = boost::accumulators::mean(accumulator);
-    auto mean_in_seconds = mean_in_milliseconds / 1000;
+    auto mean_in_seconds = mean_in_milliseconds / 1'000'000'000;
 
     if (!_config.verify && !_config.enable_visualization) {
       std::cout << "  -> Executed " << result.successful_runs.size() << " times in " << duration_seconds