This PR fixes #2226

It solves the problem the following way:

* The `benchmark_runner` now reports the latency as the arithmetic mean of iterations.
* The output latency in the benchmark_runner is now in `ms` like in the `compare_benchmarks.py`
* The order of latency and throughput are now harmonized between the live-reporting of benchmarks and the `compare_benchmarks.py` script.

Example output TPC-H 01: 

Single-Threaded:

```
- Benchmarking TPC-H 01
  -> Executed 41 times in 60.0047 seconds (Latency: 1463.1 ms/iter, Throughput: 0.68328 iter/s)
```

Multi-Threaded:
```
- Benchmarking TPC-H 01
  -> Executed 1189 times in 60.0058 seconds (Latency: 2446.3 ms/iter, Throughput: 19.8148 iter/s)
```

`compare_benchmarks.py` :

```
+----------++----------+---------+---------++----------+----------+---------+---------+
| Item     ||  Latency (ms/iter)  |  Change || Throughput (iter/s) |  Change | p-value |
|          ||      old |     new |         ||      old |      new |         |         |
+----------++----------+---------+---------++----------+----------+---------+---------+
| TPC-H 01 ||    1463.5 |  2446.7 |  +67%  ||     0.68 |    19.81 | +2800%  |  0.0000 |
```