@mrks 2021-01-03
From a code perspective, this looks great. I just ran it and the rounding discrepancies are gone.

I'd like to discuss the format though:

```
+----------++----------+---------+--------++----------+----------+--------+---------+
| Item     || Latency (ms/iter)  | Change || Throughput (iter/s) | Change | p-value |
|          ||      old |     new |        ||      old |      new |        |         |
+----------++----------+---------+--------++----------+----------+--------+---------+
| TPC-H 01 ||   1889.0 |  1868.2 |   -1%  ||    25.80 |    26.08 |   +1%  |  0.6174 |
| TPC-H 02 ||     16.4 |    16.5 |   +0%  ||  2276.52 |  2271.51 |   -0%  |  0.1964 |
| TPC-H 03 ||    205.4 |   201.0 |   -2%  ||   236.69 |   241.76 |   +2%  |  0.0699 |
| TPC-H 04 ||    155.8 |   152.8 |   -2%  ||   308.81 |   314.34 |   +2%  |  0.0867 |
| TPC-H 05 ||    542.8 |   535.2 |   -1%  ||    90.31 |    91.57 |   +1%  |  0.4951 |
| TPC-H 06 ||     10.7 |    10.2 |   -5%  ||  3210.13 |  3429.42 |   +7%  |  0.0000 |
| TPC-H 07 ||    119.9 |   117.9 |   -2%  ||   399.70 |   405.94 |   +2%  |  0.0478 |
| TPC-H 08 ||    132.9 |   130.5 |   -2%  ||   361.30 |   367.53 |   +2%  |  0.0210 |
| TPC-H 09 ||   1072.5 |  1057.0 |   -1%  ||    45.16 |    46.04 |   +2%  |  0.6064 |
| TPC-H 10 ||    281.4 |   279.5 |   -1%  ||   174.09 |   174.97 |   +1%  |  0.5162 |
| TPC-H 11 ||     33.2 |    33.4 |   +0%  ||  1323.80 |  1317.20 |   -0%  |  0.3211 |
| TPC-H 12 ||    107.2 |   105.6 |   -1%  ||   437.92 |   444.44 |   +1%  |  0.0765 |
| TPC-H 13 ||    986.4 |   990.5 |   +0%  ||    49.75 |    49.61 |   -0%  |  0.8462 |
| TPC-H 14 ||     68.1 |    67.0 |   -2%  ||   683.69 |   693.89 |   +1%  |  0.0015 |
| TPC-H 15 ||     26.5 |    26.7 |   +1%  ||  1596.94 |  1585.63 |   -1%  |  0.0060 |
| TPC-H 16 ||    223.2 |   220.6 |   -1%  ||   218.46 |   220.99 |   +1%  |  0.1948 |
| TPC-H 17 ||     38.9 |    38.3 |   -2%  ||  1125.50 |  1141.16 |   +1%  |  0.0003 |
| TPC-H 18 ||   2343.2 |  2313.7 |   -1%  ||    20.60 |    21.00 |   +2%  |  0.5761 |
| TPC-H 19 ||     71.3 |    70.7 |   -1%  ||   651.47 |   657.78 |   +1%  |  0.1279 |
| TPC-H 20 ||     44.1 |    43.5 |   -1%  ||  1035.24 |  1045.34 |   +1%  |  0.0097 |
| TPC-H 21 ||    797.4 |   780.9 |   -2%  ||    61.27 |    62.65 |   +2%  |  0.4617 |
| TPC-H 22 ||    109.7 |   107.5 |   -2%  ||   435.15 |   443.86 |   +2%  |  0.0459 |
+----------++----------+---------+--------++----------+----------+--------+---------+
| Sum      ||   9275.9 |  9167.3 |   -1%  ||          |          |        |         |
| Geomean  ||          |         |        ||          |          |   +1%  |         |
+----------++----------+---------+--------++----------+----------+--------+---------+
```

vs

```
+----------++----------+--------+--------++----------+----------+--------+---------+
| Item     || Latency (s/iter)  | Change || Throughput (iter/s) | Change | p-value |
|          ||      old |    new |        ||      old |      new |        |         |
+----------++----------+--------+--------++----------+----------+--------+---------+
| TPC-H 01 ||   1.8890 | 1.8682 |   -1%  ||    25.80 |    26.08 |   +1%  |  0.6174 |
| TPC-H 02 ||   0.0164 | 0.0165 |   +0%  ||  2276.52 |  2271.51 |   -0%  |  0.1964 |
| TPC-H 03 ||   0.2054 | 0.2010 |   -2%  ||   236.69 |   241.76 |   +2%  |  0.0699 |
| TPC-H 04 ||   0.1558 | 0.1528 |   -2%  ||   308.81 |   314.34 |   +2%  |  0.0867 |
| TPC-H 05 ||   0.5428 | 0.5352 |   -1%  ||    90.31 |    91.57 |   +1%  |  0.4951 |
| TPC-H 06 ||   0.0107 | 0.0102 |   -5%  ||  3210.13 |  3429.42 |   +7%  |  0.0000 |
| TPC-H 07 ||   0.1199 | 0.1179 |   -2%  ||   399.70 |   405.94 |   +2%  |  0.0478 |
| TPC-H 08 ||   0.1329 | 0.1305 |   -2%  ||   361.30 |   367.53 |   +2%  |  0.0210 |
| TPC-H 09 ||   1.0725 | 1.0570 |   -1%  ||    45.16 |    46.04 |   +2%  |  0.6064 |
| TPC-H 10 ||   0.2814 | 0.2795 |   -1%  ||   174.09 |   174.97 |   +1%  |  0.5162 |
| TPC-H 11 ||   0.0332 | 0.0334 |   +0%  ||  1323.80 |  1317.20 |   -0%  |  0.3211 |
| TPC-H 12 ||   0.1072 | 0.1056 |   -1%  ||   437.92 |   444.44 |   +1%  |  0.0765 |
| TPC-H 13 ||   0.9864 | 0.9905 |   +0%  ||    49.75 |    49.61 |   -0%  |  0.8462 |
| TPC-H 14 ||   0.0681 | 0.0670 |   -2%  ||   683.69 |   693.89 |   +1%  |  0.0015 |
| TPC-H 15 ||   0.0265 | 0.0267 |   +1%  ||  1596.94 |  1585.63 |   -1%  |  0.0060 |
| TPC-H 16 ||   0.2232 | 0.2206 |   -1%  ||   218.46 |   220.99 |   +1%  |  0.1948 |
| TPC-H 17 ||   0.0389 | 0.0383 |   -2%  ||  1125.50 |  1141.16 |   +1%  |  0.0003 |
| TPC-H 18 ||   2.3432 | 2.3137 |   -1%  ||    20.60 |    21.00 |   +2%  |  0.5761 |
| TPC-H 19 ||   0.0713 | 0.0707 |   -1%  ||   651.47 |   657.78 |   +1%  |  0.1279 |
| TPC-H 20 ||   0.0441 | 0.0435 |   -1%  ||  1035.24 |  1045.34 |   +1%  |  0.0097 |
| TPC-H 21 ||   0.7974 | 0.7809 |   -2%  ||    61.27 |    62.65 |   +2%  |  0.4617 |
| TPC-H 22 ||   0.1097 | 0.1075 |   -2%  ||   435.15 |   443.86 |   +2%  |  0.0459 |
+----------++----------+--------+--------++----------+----------+--------+---------+
| Sum      ||   9.2759 | 9.1673 |   -1%  ||          |          |        |         |
| Geomean  ||          |        |        ||          |          |   +1%  |         |
+----------++----------+--------+--------++----------+----------+--------+---------+
```

I find the latency column in the latter example harder to read. Personally, I would also be fine with `ms/iter` and `iter/s`. After all, all SQL queries should be subsecond, so why would we need `s/iter` anyway? ;)

No strong opinion about this though. If you and @Bouncner prefer this version, I'm fine with it, too.