@Bouncner 2021-01-03
I am not a big fan of the inconsistency with `ms/iter` and `iter/s`, BUT since I read the results of the compare script so often, I rather have a nice to read output. So I am also in favor of having `ms/iter` and `iter/s`.