@Alexander-Dubrawski 2021-01-04
> I am not a big fan of the inconsistency with `ms/iter` and `iter/s`, BUT since I read the results of the compare script so often, I rather have a nice to read output. So I am also in favor of having `ms/iter` and `iter/s`.

Okay, I will change it back. Whats about the latency output in the `benchmark_runner`, should it be in `ms` as well?