@mrks 2021-01-09
Let me know if you are stuck with the tsan issue and we can look at it together.