@Alexander-Dubrawski 2021-01-14
> Let me know if you are stuck with the tsan issue and we can look at it together.

I think I need some pinpoints. I thought it could be a synchronization problem. Since some workers are still doing some work (writing some `BenchmarkItemRunResult` in `result.successful_runs`). But `result.successful_runs` is a `tbb::concurrent_vector` so there shouldn't be a synchronization problem, or am I missing somethink?