@mrks 2021-01-14
> But result.successful_runs is a tbb::concurrent_vector so there shouldn't be a synchronization problem, or am I missing something?

That's partially correct and I was living under that impression for too long myself. A concurrent_vector that does guarantee that the objects inside it are read/written atomically, it just provides atomic container methods. Have a look at [this post](http://web.archive.org/web/20100103204609/http://software.intel.com/en-us/blogs/2009/04/09/delusion-of-tbbconcurrent_vectors-size-or-3-ways-to-traverse-in-parallel-correctly/).

> The result may include elements that are under construction by concurrent calls to any of the growth methods

If my guess is correct and this is the cause for tsan complaining, it could be simply solved by waiting for all clients to finish (even the ones that will not be counted anymore). We should probably print something like "waiting for clients that are still running" so that no one wonders why a benchmark that is supposed to take 30s runs for longer.